pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "golf-swing-analyzer"
version = "0.1.0"
description = "ESP32, Raspberry Pi, and IMU sensor IoT system for golf swing motion analysis and ball trajectory prediction"
authors = [
    {name = "Golf Swing Analyzer Team"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "bleak>=0.21.1",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "asyncio-mqtt>=0.13.0",
    "websockets>=11.0.3",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.4.0",
    "matplotlib>=3.7.0",
    "pandas>=2.0.0",
    "pyyaml>=6.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "pre-commit>=3.3.0",
]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true